Worked on the charts on the professor gradebook. Fix random bugs

+++ blossom/controllers/RestfulApi.php
@@ -84,7 +84,7 @@ class RestfulApi extends Controller
         
         foreach($this->submissions as $item)
         {
-            $score = $score+intval($item['score']);
+            $score = $score+floatval($item['score']);
         }
         return $score;
     }
